+++ packages/f-checkout/src/components/tests/Checkout.test.js
@@ -1,10 +1,21 @@
-import { shallowMount, mount } from '@vue/test-utils';
+import { shallowMount, mount, createLocalVue } from '@vue/test-utils';
+import { VueI18n } from '@justeat/f-globalisation';
 import { VALID_CHECKOUT_METHOD } from '../../constants';
 import VueCheckout from '../Checkout.vue';
 import CheckoutServiceApi from '../../services/CheckoutServiceApi';
 import EventNames from '../../event-names';
+import tenantConfigs from '../../tenants';
 
 jest.mock('../../services/CheckoutServiceApi', () => ({ submitCheckout: jest.fn() }));
+
+const localVue = createLocalVue();
+
+localVue.use(VueI18n);
+
+const i18n = {
+    locale: 'en-GB',
+    messages: tenantConfigs['en-GB']
+};
 
 describe('Checkout', () => {
     allure.feature('Checkout');
@@ -13,7 +24,13 @@ describe('Checkout', () => {
 
     it('should be defined', () => {
         const propsData = { checkoutUrl };
-        const wrapper = shallowMount(VueCheckout, { propsData });
+
+        const wrapper = shallowMount(VueCheckout, {
+            i18n,
+            localVue,
+            propsData
+        });
+
         expect(wrapper.exists()).toBe(true);
     });
 
@@ -27,7 +44,12 @@ describe('Checkout', () => {
                 };
 
                 // Act
-                const wrapper = shallowMount(VueCheckout, { propsData });
+                const wrapper = shallowMount(VueCheckout, {
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
                 const selectorComponent = wrapper.find('[data-test-id="selector"]');
 
                 // Assert
@@ -42,7 +64,12 @@ describe('Checkout', () => {
                 };
 
                 // Act
-                const wrapper = shallowMount(VueCheckout, { propsData });
+                const wrapper = shallowMount(VueCheckout, {
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
                 const addressBlock = wrapper.find('[data-test-id="address-block"]');
 
                 // Assert
@@ -57,7 +84,12 @@ describe('Checkout', () => {
                 };
 
                 // Act
-                const wrapper = shallowMount(VueCheckout, { propsData });
+                const wrapper = shallowMount(VueCheckout, {
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
                 const addressBlock = wrapper.find('[data-test-id="address-block"]');
 
                 // Assert
@@ -73,7 +105,12 @@ describe('Checkout', () => {
         describe('name ::', () => {
             it('should capitalize `firstName` data', async () => {
                 // Arrange
-                const wrapper = shallowMount(VueCheckout, { propsData });
+                const wrapper = shallowMount(VueCheckout, {
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
                 const name = wrapper.find("[data-test-id='checkout-card-component']");
 
                 // Act
@@ -88,7 +125,12 @@ describe('Checkout', () => {
         describe('title ::', () => {
             it('should add `name` to title text', async () => {
                 // Arrange
-                const wrapper = shallowMount(VueCheckout, { propsData });
+                const wrapper = shallowMount(VueCheckout, {
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
                 const name = wrapper.find("[data-test-id='checkout-card-component']");
 
                 // Act