+++ packages/f-checkout/src/tenants/en-GB.js
@@ -25,10 +25,12 @@ export default {
         }
     },
 
-    allergyText: 'If you or someone you’re ordering for has a food allergy or intolerance, click here',
-
     errorMessages: {
         errorHeading: 'Error',
         genericServerError: 'Something went wrong, please try again later'
-    }
+    },
+  
+    buttonText: 'Go to payment',
+
+    allergyText: 'If you or someone you’re ordering for has a food allergy or intolerance, click here'
 };