+++ packages/f-checkout/CHANGELOG.md
@@ -3,15 +3,25 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.8.0
+------------------------------
+*November 12, 2020*
+
+### Added
+- Inline error handling to `Checkout` and `Address` components.
+- Server side error handling.
+- Validations from `f-services`.
+
 
 v0.7.0
 ------------------------------
 *November 11, 2020*
 
 ### Added
-- Inline error handling to `Checkout` and `Address` components.
-- Server side error handling.
-- Validations from `f-services`.
+- Integration with `f-globalisation`.
+
+### Changed
+- The way we run the demo page (used by component tests) so that we can control what we inject into the Vue instance.
 
 
 v0.6.0